+++ examples/configuration_archive/main.go
@@ -15,7 +15,7 @@ func main() {
 	deviceUUID := ""
 	client, err = dnac.NewClientWithOptions("https://192.168.196.2/",
 		"altus", "Altus123",
-		"true", "false")
+		"true", "false", nil)
 	if err != nil {
 		fmt.Println(err)
 		return