+++ examples/swim/main.go
@@ -16,7 +16,7 @@ func main() {
 	fmt.Println("Authenticating")
 	client, err = dnac.NewClientWithOptions("https://192.168.196.2/",
 		"altus", "Altus123",
-		"false", "false")
+		"false", "false", nil)
 	if err != nil {
 		fmt.Println(err)
 		return